move eventlistner to correct package; update implementation
