+++ airbyte-server/src/main/java/io/airbyte/server/DatabaseEventListener.java
@@ -2,13 +2,13 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.server.services;
+package io.airbyte.server;
 
 import io.airbyte.db.check.DatabaseCheckException;
 import io.airbyte.db.check.DatabaseMigrationCheck;
-import io.micronaut.context.event.ApplicationEventListener;
+import io.micronaut.context.event.StartupEvent;
 import io.micronaut.core.annotation.Order;
-import io.micronaut.runtime.server.event.ServerStartupEvent;
+import io.micronaut.runtime.event.annotation.EventListener;
 import jakarta.inject.Named;
 import jakarta.inject.Singleton;
 import java.lang.invoke.MethodHandles;
@@ -16,8 +16,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Singleton
-@Order(1)
-public class DatabaseEventListener implements ApplicationEventListener<ServerStartupEvent> {
+public class DatabaseEventListener {
 
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
@@ -32,8 +31,9 @@ public class DatabaseEventListener implements ApplicationEventListener<ServerSta
     this.jobsMigrationCheck = jobsMigrationCheck;
   }
 
-  @Override
-  public void onApplicationEvent(final ServerStartupEvent event) {
+  @EventListener
+  @Order(1)
+  public void onStartup(final StartupEvent event) {
     log.info("Checking configs database flyway migration version...");
     try {
       configsMigrationCheck.check();
